requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
addict=2.4.0=pypi_0
altair=4.1.0=py_1
arrow-cpp=8.0.0=py310h38b8b19_1
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py310haa95532_0
aws-c-common=0.4.57=ha925a31_1
aws-c-event-stream=0.1.6=hd77b12b_5
aws-checksums=0.1.9=ha925a31_0
aws-sdk-cpp=1.8.185=hd77b12b_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py310haa95532_0
boost-cpp=1.73.0=h2bbff1b_12
bottleneck=1.3.5=py310h9128911_0
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
c-ares=1.19.0=h2bbff1b_0
ca-certificates=2023.01.10=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=pypi_0
colorama=0.4.6=py310haa95532_0
comm=0.1.2=py310haa95532_0
commonmark=0.9.1=pyhd3eb1b0_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py310h21b164f_0
cuda-cccl=12.1.55=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.1.55=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pypi_0
cython=0.29.33=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
eigen=3.3.7=h59b6b97_1
entrypoints=0.4=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.2.2=he774522_0
filelock=3.9.0=py310haa95532_0
flit-core=3.8.0=py310haa95532_0
fonttools=4.39.2=pypi_0
freetype=2.12.1=ha860e81_0
future=0.18.3=py310haa95532_0
gflags=2.2.2=ha925a31_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py310haa95532_0
glib=2.69.1=h5dc1a3c_2
glog=0.5.0=hd77b12b_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.10.6=h1756f20_1
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=py310haa95532_0
importlib-metadata=6.0.0=py310haa95532_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.19.2=py310h9909e9c_0
ipython=8.10.0=py310haa95532_0
ipywidgets=8.0.4=py310haa95532_0
jedi=0.18.1=py310haa95532_1
jinja2=3.1.2=py310haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=py310haa95532_0
jupyter_client=7.4.9=py310haa95532_0
jupyter_core=5.2.0=py310haa95532_0
jupyterlab_widgets=3.0.5=py310haa95532_0
kiwisolver=1.4.4=pypi_0
lerc=3.0=hd77b12b_0
libboost=1.73.0=h6c2663c_12
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.2.56=0
libcurand-dev=10.3.2.56=0
libcurl=7.88.1=h86230a5_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libogg=1.3.5=h2bbff1b_1
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=hcd4344a_0
libthrift=0.15.0=he1d8c1a_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.4.3=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.1=py310h2bbff1b_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=py310haa95532_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
mmcls=0.25.0=pypi_0
mmcv=1.7.1=pypi_0
mmcv-full=1.7.1=pypi_0
mmdet=2.28.1=dev_0
mmpycocotools=12.0.3=pypi_0
model-index=0.1.11=pypi_0
mpmath=1.2.1=py310haa95532_0
nest-asyncio=1.5.6=py310haa95532_0
networkx=2.8.4=py310haa95532_1
numexpr=2.8.4=py310hd213c9f_0
numpy=1.23.5=py310h60c9a35_0
numpy-base=1.23.5=py310h04254f7_0
opencv=4.6.0=py310h4ed8f06_3
openmim=0.3.7=pypi_0
openssl=1.1.1t=h2bbff1b_0
ordered-set=4.1.0=pypi_0
packaging=23.0=py310haa95532_0
pandas=1.5.3=py310h4ed8f06_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py310hd77b12b_0
pip=23.0.1=py310haa95532_0
platformdirs=2.5.2=py310haa95532_0
prompt-toolkit=3.0.36=py310haa95532_0
protobuf=3.20.3=py310hd77b12b_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=py310h26aae1b_0
pycocotools=2.0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py310haa95532_0
pygments=2.14.0=pypi_0
pympler=0.9=py_0
pyopenssl=23.0.0=py310haa95532_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.0=py310h2bbff1b_0
pysocks=1.7.1=py310haa95532_0
python=3.10.10=h966fe2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=2.0.0=py3.10_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_3
pytorch-mutex=1.0=cuda
pytz=2023.2=pypi_0
pywin32=305=py310h2bbff1b_0
pyyaml=6.0=py310h2bbff1b_1
pyzmq=23.2.0=py310hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
re2=2022.04.01=hd77b12b_0
regex=2023.3.23=pypi_0
requests=2.28.1=py310haa95532_1
rich=13.3.2=pypi_0
scipy=1.10.1=pypi_0
semver=2.13.0=pyhd3eb1b0_0
setuptools=65.6.3=py310haa95532_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h6c2663c_0
sqlite=3.41.1=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.20.0=pypi_0
sympy=1.11.1=py310haa95532_0
tabulate=0.9.0=pypi_0
terminaltables=3.1.10=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py310haa95532_0
torchaudio=2.0.0=pypi_0
torchvision=0.15.0=pypi_0
tornado=6.2=py310h2bbff1b_0
traitlets=5.7.1=py310haa95532_0
typing_extensions=4.4.0=py310haa95532_0
tzdata=2022g=h04d1e81_0
tzlocal=2.1=py310haa95532_1
urllib3=1.26.14=py310haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.18.2=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py310haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py310haa95532_0
widgetsnbextension=4.0.5=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xz=5.2.10=h8cc25b3_1
yaml=0.2.5=he774522_0
yapf=0.32.0=pypi_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py310haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
